more explicit names for BalanceService functions
